refactor: streamline BookingService constructor and standardize property names across services

+++ backend/Backend.Services/Validators/Hall/UpdateHallDtoValidator.cs
@@ -12,6 +12,6 @@ public class UpdateHallDtoValidator : AbstractValidator<UpdateHallDto>
             .MaximumLength(50).WithMessage("Назва не може бути довшою за 50 символів");
 
         RuleFor(x => x.Format)
-            .InclusiveBetween((short)0, (short)5).WithMessage("Некоректний формат залу");
+            .InclusiveBetween(0, 5).WithMessage("Некоректний формат залу");
     }
 }
